[#395] CHORE: component명 수정
